% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_years}
\alias{available_years}
\title{List Avialable Years Based on Meta Information}
\usage{
available_years(collection_root)
}
\arguments{
\item{collection_root}{character file path to collection.}
}
\description{
Checks which yearly collections are available based on
meta information file names.
}
