% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_counting.R
\name{average_length_by_date}
\alias{average_length_by_date}
\title{Average Ad length by Date}
\usage{
average_length_by_date(coll, ids = NULL, level = NULL, unit = "tokens")
}
\arguments{
\item{coll}{object of class Collection}

\item{ids}{character vector of ids to filter for. Defaults to NULL (all).}

\item{level}{character year, quarter, month, week. Defaults to NULL (average).}

\item{unit}{character denoting 'tokens' or 'character'}
}
\description{
Average Ad length by Date
}
