% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avis_create_corpus.R
\name{avis_create_corpus}
\alias{avis_create_corpus}
\title{Reads Avisblatt Input CSV into Quantedata Corpus}
\usage{
avis_create_corpus(
  f,
  text_field = "text",
  docid_field = "id",
  encoding = "UTF-8",
  apply_ocr_correct = FALSE
)
}
\arguments{
\item{f}{character location of the file to read.}

\item{text_field}{name of the text column.}

\item{docid_field}{name of the document id field.}

\item{encoding}{character defaults to UTF-8.}

\item{apply_ocr_correct}{boolean Defaults to FALSE.}
}
\description{
Convenience function to read avisblatt input .csvs into
quanteda 2.0 corpuses.
}
