% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avis_stop.R
\name{avis_stop}
\alias{avis_stop}
\title{Avisblatt Specific Stopword Dictionary}
\usage{
avis_stop(s = NULL, keep_standard = TRUE)
}
\arguments{
\item{s}{character vector of stopwords}

\item{keep_standard}{boolean should default stopwords be kept when new stopwords are provided? Defaulst to TRUE.}
}
\value{
list containing vectors of stopwords.
}
\description{
Collection of stopwords.
}
