% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avisblatt.R
\docType{package}
\name{avisblatt-package}
\alias{avisblatt-package}
\alias{avisblatt}
\title{Process Text from Early Modern Source Avisblatt}
\description{
Process Text From the Early Modern Avisblatt Source
}
\examples{
\dontrun{
demo("echo", package="avisblatt")
}


}
\author{
Matthias Bannert \email{bunny@hey.com}
}
\keyword{package}
