% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_manual_tags}
\alias{clean_manual_tags}
\title{Clean up Human Tags (groundtruth)}
\usage{
clean_manual_tags(x)
}
\arguments{
\item{x}{character vector or list of tags.}
}
\description{
Removes leading numbers from tag and turn comma separated strings
into true character vectors.
}
