% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{Collection}
\alias{Collection}
\title{R6 Class Representing a Collection}
\description{
Avisblatt collection of data and meta data.
}
\details{
Collections can be written to disk in two linked files: csv and json.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{corpus}}{object of class Corpus}

\item{\code{meta}}{object of class meta}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Collection-new}{\code{Collection$new()}}
\item \href{#method-Collection-add_tags}{\code{Collection$add_tags()}}
\item \href{#method-Collection-add_language}{\code{Collection$add_language()}}
\item \href{#method-Collection-list_records}{\code{Collection$list_records()}}
\item \href{#method-Collection-apply_tagfilters}{\code{Collection$apply_tagfilters()}}
\item \href{#method-Collection-subset_collect}{\code{Collection$subset_collect()}}
\item \href{#method-Collection-remove_records}{\code{Collection$remove_records()}}
\item \href{#method-Collection-clone}{\code{Collection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collection-new"></a>}}
\if{latex}{\out{\hypertarget{method-Collection-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$new(
  crps,
  meta_table = NULL,
  preprocess_docvars = NULL,
  docvars_to_meta = NULL,
  transform_docvars = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{crps}}{object of class quanteda corpus.}

\item{\code{meta_table}}{meta information data.table}

\item{\code{preprocess_docvars}}{docvars}

\item{\code{docvars_to_meta}}{docvars to meta}

\item{\code{transform_docvars}}{transform docvars}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collection-add_tags"></a>}}
\if{latex}{\out{\hypertarget{method-Collection-add_tags}{}}}
\subsection{Method \code{add_tags()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$add_tags(ids = NULL, tags_vec, overwrite = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{character vector of document ids}

\item{\code{tags_vec}}{add vector of tags}

\item{\code{overwrite}}{boolean should existing tags be overwritten. Defaults to FALSE}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collection-add_language"></a>}}
\if{latex}{\out{\hypertarget{method-Collection-add_language}{}}}
\subsection{Method \code{add_language()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$add_language(ids = NULL, lang)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{character vector of ids}

\item{\code{lang}}{character 2-digit ISO language}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collection-list_records"></a>}}
\if{latex}{\out{\hypertarget{method-Collection-list_records}{}}}
\subsection{Method \code{list_records()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$list_records(ids = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{character vector of ids. Defaults to NULL, selecting all.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collection-apply_tagfilters"></a>}}
\if{latex}{\out{\hypertarget{method-Collection-apply_tagfilters}{}}}
\subsection{Method \code{apply_tagfilters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$apply_tagfilters(flist, nms = NULL, ignore_case = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{flist}}{list of filters}

\item{\code{nms}}{character names}

\item{\code{ignore_case}}{boolean Should case be ignored? Defaults to FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collection-subset_collect"></a>}}
\if{latex}{\out{\hypertarget{method-Collection-subset_collect}{}}}
\subsection{Method \code{subset_collect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$subset_collect(ids)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{character list of ids}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collection-remove_records"></a>}}
\if{latex}{\out{\hypertarget{method-Collection-remove_records}{}}}
\subsection{Method \code{remove_records()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$remove_records(ids, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{character vector of ids}

\item{\code{verbose}}{boolean should things be verbose? Defaults to TRUE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Collection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
