% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocr_corrections.R
\name{correct_ocr}
\alias{correct_ocr}
\title{Correction of OCR Issues}
\usage{
correct_ocr(x)
}
\arguments{
\item{x}{character vector to be corrected}
}
\value{
character vector
}
\description{
Regular expression based correction of scanning issues.
}
