% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections.R
\docType{data}
\name{corrections}
\alias{corrections}
\title{Corrections Based on Manual Work}
\description{
Various overrides of automatically attached meta information based on
manual work.
}
\keyword{data}
