% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_counting.R
\name{count_records_by_date}
\alias{count_records_by_date}
\title{Count records by date}
\usage{
count_records_by_date(
  coll,
  ids = NULL,
  level = "year",
  colnames = NULL,
  per1k = FALSE,
  trim = TRUE,
  min = "1729-01-01",
  max = "1844-12-31",
  bundle_periods = 1
)
}
\arguments{
\item{coll}{A collection object}

\item{ids}{A character vector of ids to count records for}

\item{level}{The level at which to count records (year, quarter, month, week, issue)}

\item{colnames}{A character vector of column names for the output data.table}

\item{per1k}{Logical indicating whether to return counts per 1000 records}

\item{trim}{Logical indicating whether to trim the first and last bundle_periods periods from the output data.table}

\item{min}{The minimum date to include in the output data.table}

\item{max}{The maximum date to include in the output data.table}

\item{bundle_periods}{The number of periods to bundle together when counting records (e.g. 3 months)}
}
\value{
A data.table with counts of records by date and id(s)
}
\description{
Count records by date
}
