% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_counting.R
\name{count_records_by_length}
\alias{count_records_by_length}
\title{Records by Length}
\usage{
count_records_by_length(
  coll,
  ids = NULL,
  boundaries = c(0, 10, 20, 40, 80, 160, 1000),
  unit = "tokens"
)
}
\arguments{
\item{coll}{object of class Collection.}

\item{ids}{character vector of ids to filter for. Defaults to NULL (all).}

\item{boundaries}{numeric boundaries of length bins}

\item{unit}{character denoting 'tokens' or 'character'}
}
\description{
Count the number of ads by length bins. Bins can be character or token
based
}
