% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaners.R
\name{detect_lang}
\alias{detect_lang}
\title{Return all Document IDs of a Particular Language}
\usage{
detect_lang(corp, lang = "french")
}
\arguments{
\item{corp}{quanteda corpus object}

\item{lang}{character language to detect, defaults to french.}
}
\value{
document ids of language
}
\description{
Based in the textcat package, this function tries to detect a
given language in a document corpus and returns all document ids of
documents written in the desired language. While German is hard to
detect for in these historical texts, it does a very good job
seperating French and German when looking for French texts.
}
