% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gather_yearly_collections}
\alias{gather_yearly_collections}
\title{Convenience Function to Read & Merge Multiple Years}
\usage{
gather_yearly_collections(
  AVIS_YEARS,
  just_meta = TRUE,
  path = "../avis-data/collections"
)
}
\arguments{
\item{AVIS_YEARS}{numeric vectors of selected years.}

\item{just_meta}{boolean should meta be used w/o reading in the actual text?
Defaults to TRUE.}

\item{path}{character location of the collection files.}
}
\description{
Convenience Function to Read & Merge Multiple Years
}
