% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_scouting.R
\name{get_headers}
\alias{get_headers}
\title{Get Headers from Collection}
\usage{
get_headers(coll, text = FALSE)
}
\arguments{
\item{coll}{object of class Collection}

\item{text}{boolean. Defaults to FALSE.}
}
\description{
Get Headers from Collection
}
