% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_scouting.R
\name{get_noadverts}
\alias{get_noadverts}
\title{Get noadvert records from a Collection object
This function returns the noadvert records from a Collection object.}
\usage{
get_noadverts(coll, text = FALSE)
}
\arguments{
\item{coll}{A Collection object.}

\item{text}{A logical value indicating whether to return the text of the noadvert records or their names.}
}
\value{
If text is TRUE, the text of the noadvert records; otherwise, their names.
}
\description{
Get noadvert records from a Collection object
This function returns the noadvert records from a Collection object.
}
\seealso{
\code{\link{read_collection}}
}
\keyword{collection}
