% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_text_by_id}
\alias{get_text_by_id}
\title{Might want to deprecate this with quanteda 2.0}
\usage{
get_text_by_id(corp, ids, n = NULL, identifier = "id", txt = "texts")
}
\description{
Might want to deprecate this with quanteda 2.0
}
