% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_grouping.R
\name{group_records_by_header}
\alias{group_records_by_header}
\title{Group Records by Header}
\usage{
group_records_by_header(coll, ids = NULL, headerlist = NULL, namelist = NULL)
}
\arguments{
\item{coll}{object of class Collection.}

\item{ids}{character vector of ids to filter for. Defaults to NULL (all).}
}
\description{
Group Records by Header
}
