% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_grouping.R
\name{group_records_by_tags}
\alias{group_records_by_tags}
\title{Group Records by Tags}
\usage{
group_records_by_tags(coll, ids = NULL, tagslist = NULL, namelist = NULL)
}
\arguments{
\item{coll}{object of class Collection.}

\item{ids}{character vector of ids to filter for. Defaults to NULL (all).}

\item{tagslist}{list of tags}

\item{namelist}{list of names}
}
\description{
Group Records by Tags
}
