% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_grouping.R
\name{group_records_by_text}
\alias{group_records_by_text}
\title{Group Records by Text}
\usage{
group_records_by_text(coll, ids = NULL, searchlist = "", namelist = NULL)
}
\arguments{
\item{coll}{object of class Collection.}

\item{ids}{character vector of ids to filter for. Defaults to NULL (all).}

\item{searchlist}{character defaults to empty.}

\item{namelist}{defaults to NULL.}
}
\description{
Group Records by Text
}
