% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_scouting.R
\name{list_similar}
\alias{list_similar}
\title{List similar ads}
\usage{
list_similar(
  coll,
  ids1 = NULL,
  ids2 = NULL,
  N = 5,
  min_dist = 0,
  max_dist = 1000,
  matrix_limit = 1e+07
)
}
\arguments{
\item{coll}{A collection object.}

\item{ids1}{A vector of ad IDs from the first collection.}

\item{ids2}{A vector of ad IDs from the second collection.}

\item{N}{The number of most similar ads to return.}

\item{min_dist}{The minimum distance between ads to consider.}

\item{max_dist}{The maximum distance between ads to consider.}

\item{matrix_limit}{The maximum number of elements in the distance matrix.}
}
\value{
A matrix with the N most similar ads and their distances.
}
\description{
This function calculates the distance between two collections of ad texts and returns the N most similar ads.
}
\examples{
\dontrun{
list_similar(coll = coll,
             ids1 = c("ad1", "ad2", "ad3"),
             ids2 = c("ad4", "ad5", "ad6"),
             N = 5,
             min_dist = 0,
             max_dist = 1000,
             matrix_limit = 10000000)
}

}
