% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdata.R
\name{RawData}
\alias{RawData}
\title{R6 Class Representing Raw Data}
\description{
Avisblatt raw data as pulled from FREIZO archive where OCR'ed
avisblatt information is.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{two dimensional dataset}

\item{\code{year}}{numeric year}

\item{\code{status}}{character status description}

\item{\code{message}}{character message}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RawData-new}{\code{RawData$new()}}
\item \href{#method-RawData-booleanize}{\code{RawData$booleanize()}}
\item \href{#method-RawData-drop_on_condition}{\code{RawData$drop_on_condition()}}
\item \href{#method-RawData-drop_cols}{\code{RawData$drop_cols()}}
\item \href{#method-RawData-process_data}{\code{RawData$process_data()}}
\item \href{#method-RawData-write_log_status}{\code{RawData$write_log_status()}}
\item \href{#method-RawData-write_csv}{\code{RawData$write_csv()}}
\item \href{#method-RawData-create_header_tags}{\code{RawData$create_header_tags()}}
\item \href{#method-RawData-clone}{\code{RawData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawData-new"></a>}}
\if{latex}{\out{\hypertarget{method-RawData-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawData$new(
  year,
  URL = "https://avisblatt.freizo.org/curator/annotsv.cgi?year=",
  local = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{year}}{numeric year}

\item{\code{URL}}{character URL to FREIZO archive}

\item{\code{local}}{boolean}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawData-booleanize"></a>}}
\if{latex}{\out{\hypertarget{method-RawData-booleanize}{}}}
\subsection{Method \code{booleanize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawData$booleanize(cols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cols}}{character names of coiumns to put into booleans}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawData-drop_on_condition"></a>}}
\if{latex}{\out{\hypertarget{method-RawData-drop_on_condition}{}}}
\subsection{Method \code{drop_on_condition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawData$drop_on_condition(col, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col}}{column to drop based on a condition}

\item{\code{value}}{conditional value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawData-drop_cols"></a>}}
\if{latex}{\out{\hypertarget{method-RawData-drop_cols}{}}}
\subsection{Method \code{drop_cols()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawData$drop_cols(col)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col}}{columns to drop}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawData-process_data"></a>}}
\if{latex}{\out{\hypertarget{method-RawData-process_data}{}}}
\subsection{Method \code{process_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawData$process_data(
  clean_formatting = TRUE,
  ocr_correction = TRUE,
  remove_test_data = TRUE,
  remove_invalid_data = TRUE,
  remove_duplicates = TRUE,
  drop_na_cols = TRUE,
  gt_years = c(1734, 1754, 1774, 1834),
  test_users = c("Dan Granville", "Peter Cornwell")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clean_formatting}}{boolean clean formatting?}

\item{\code{ocr_correction}}{boolean ocr correction?}

\item{\code{remove_test_data}}{boolean remove test data?}

\item{\code{remove_invalid_data}}{boolean remove invalid data ?}

\item{\code{remove_duplicates}}{boolean remove duplicates?}

\item{\code{drop_na_cols}}{boolean NA columns, defaults to TRUE}

\item{\code{gt_years}}{numeric vector of ground truth years}

\item{\code{test_users}}{character vector of test user names}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawData-write_log_status"></a>}}
\if{latex}{\out{\hypertarget{method-RawData-write_log_status}{}}}
\subsection{Method \code{write_log_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawData$write_log_status(
  log_file = "README.md",
  mark_down_table_log_format = TRUE,
  line_end = "\\n"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log_file}}{character file log entry destination}

\item{\code{mark_down_table_log_format}}{boolean}

\item{\code{line_end}}{character to determine line endings.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawData-write_csv"></a>}}
\if{latex}{\out{\hypertarget{method-RawData-write_csv}{}}}
\subsection{Method \code{write_csv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawData$write_csv(fn = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fn}}{character file name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawData-create_header_tags"></a>}}
\if{latex}{\out{\hypertarget{method-RawData-create_header_tags}{}}}
\subsection{Method \code{create_header_tags()}}{
Create Tags Based on Section Headers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawData$create_header_tags()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RawData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
