% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdata_processing.R
\name{rawdata_apply_ocr}
\alias{rawdata_apply_ocr}
\title{Apply OCR to Raw Data}
\usage{
rawdata_apply_ocr(
  AVIS_YEARS = 1729:1844,
  source_path = "../avis-databuffer/raw_data_uncorrected",
  dest_path = "../avis-databuffer/raw_data_OCRed"
)
}
\arguments{
\item{AVIS_YEARS}{numeric vector of selected years.}

\item{source_path}{character path to collection}

\item{dest_path}{character path to destination}
}
\description{
Apply OCR to Raw Data
}
