% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdata_processing.R
\name{rawdata_coll_creation}
\alias{rawdata_coll_creation}
\title{Create Raw Data Collection}
\usage{
rawdata_coll_creation(
  AVIS_YEARS = 1729:1844,
  source_path = "../avis-data/raw_data",
  dest_path = "../avis-databuffer/collections_unflagged",
  gt_years = c(1734, 1754, 1774, 1834)
)
}
\arguments{
\item{AVIS_YEARS}{numeric vector of selected years.}

\item{source_path}{character path to collection}

\item{dest_path}{character path to destination}

\item{gt_years}{numeric vector of ground truth years.}
}
\description{
Create Raw Data Collection
}
