% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdata_flagging.R
\name{rawdata_fraternaltwin_detection}
\alias{rawdata_fraternaltwin_detection}
\title{Find Fraternal Twins}
\usage{
rawdata_fraternaltwin_detection(
  AVIS_YEARS = 1729:1844,
  source_path_csv = "../avis-databuffer/collections_unflagged",
  source_path_json = "../avis-databuffer/collections_reprint_detected",
  dest_path = "../avis-data/collections"
)
}
\arguments{
\item{AVIS_YEARS}{numeric vector of selected years.}

\item{source_path_csv}{character path to csv}

\item{source_path_json}{character path to json}

\item{dest_path}{character path to destination}
}
\description{
Find Fraternal Twins
}
