% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdata_processing.R
\name{rawdata_header_and_id}
\alias{rawdata_header_and_id}
\title{Rawdata Header and ID}
\usage{
rawdata_header_and_id(
  AVIS_YEARS = 1729:1844,
  source_path = "../avis-databuffer/raw_data_OCRed",
  dest_path = "../avis-data/raw_data"
)
}
\arguments{
\item{AVIS_YEARS}{numeric vector of selected years.}

\item{source_path}{character path to collection}

\item{dest_path}{character path to destination}
}
\description{
Rawdata Header and ID
}
