% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdata_flagging.R
\name{rawdata_redo_tags}
\alias{rawdata_redo_tags}
\title{Redo Tags}
\usage{
rawdata_redo_tags(
  AVIS_YEARS = 1729:1844,
  path = "../avis-data/collections",
  path_no_ft = "../avis-databuffer/collections_reprint_detected",
  path_no_rp = "../avis-databuffer/collections_unflagged",
  correct_final = T,
  correct_no_ft = T,
  correct_no_rp = T
)
}
\arguments{
\item{AVIS_YEARS}{numeric vector of selected years.}

\item{path}{character path to collections}

\item{path_no_ft}{character path no_ft}

\item{path_no_rp}{character path no_rp}

\item{correct_final}{boolean should final correction be applied, defaults to TRUE.}

\item{correct_no_ft}{boolean final no ft correction? defaults to TRUE.}

\item{correct_no_rp}{boolean final no rp correction? defaults to TRUE.}
}
\description{
Redo Tags
}
