% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdata_flagging.R
\name{rawdata_reprint_detection}
\alias{rawdata_reprint_detection}
\title{Detect Reprint}
\usage{
rawdata_reprint_detection(
  AVIS_YEARS = 1729:1844,
  source_path = "../avis-databuffer/collections_unflagged",
  dest_path = "../avis-databuffer/collections_reprint_detected"
)
}
\arguments{
\item{AVIS_YEARS}{numeric vector of selected years.}

\item{source_path}{character path to collection}

\item{dest_path}{character path to destination}
}
\description{
Detect Reprint
}
