% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_collection.R
\name{read_collection}
\alias{read_collection}
\title{Read Avisblatt Collection of Data and Metadata}
\usage{
read_collection(name_on_disk, just_meta = FALSE)
}
\description{
In a separation of concerns approach the avisblatt project splits data and
data descriptions into two files best suited for the different nature of the
information represented. Doing so makes the entire information machine
friendly as the split allows to use standard read/write functions for .csv
and .json file formats.
}
