% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_selecting.R
\name{select_by_date}
\alias{select_by_date}
\title{Selects data from a collection object based on date range}
\usage{
select_by_date(coll, ids = NULL, min = "1729-01-01", max = "1844-12-31")
}
\arguments{
\item{coll}{A Collection object.}

\item{ids}{A character vector of ids to select from. If NULL, all ids are selected.}

\item{min}{A character string representing the minimum date in format "YYYY-MM-DD".}

\item{max}{A character string representing the maximum date in format "YYYY-MM-DD".}
}
\value{
A character vector of ids that meet the date range criteria.
}
\description{
Selects data from a collection object based on date range
}
