% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_selecting.R
\name{select_by_language}
\alias{select_by_language}
\title{Selects documents by language}
\usage{
select_by_language(coll, ids = NULL, status = c("de", "fr", "unknown"))
}
\arguments{
\item{coll}{A collection object.}

\item{ids}{A vector of document IDs to select. If NULL, all documents are selected.}

\item{status}{A character string indicating the language status of the documents to select. Possible values are "de", "fr", and "unknown".}
}
\value{
A vector of document IDs that match the specified criteria.
}
\description{
This function selects documents from a collection based on their language.
}
