% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_selecting.R
\name{select_by_meta}
\alias{select_by_meta}
\title{Selects ad by meta information}
\usage{
select_by_meta(coll, ids = NULL, search = NULL, fields = NULL, exact = FALSE)
}
\arguments{
\item{coll}{an object of class Collection.}

\item{ids}{character vector of ids to be searched. Defaults to NULL, searching all
available character.}

\item{search}{character what look for}

\item{fields}{character where to search}

\item{exact}{boolean should matches be exact? Defaults to FALSE.}
}
\description{
Selects ad by meta information
}
