% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_selecting.R
\name{select_by_reprint_status}
\alias{select_by_reprint_status}
\title{Selects records from a collection based on reprint status}
\usage{
select_by_reprint_status(
  coll,
  ids = NULL,
  status = c("reprinted_orig", "unreprinted_orig", "postings", "reprints", "other",
    "ads")
)
}
\arguments{
\item{coll}{A collection object.}

\item{ids}{A vector of record ids to select. If NULL (default), all records are selected.}

\item{status}{A character string specifying the reprint status to select. Possible values are:
\itemize{
  \item "reprinted_orig": Reprinted originals
  \item "unreprinted_orig": Unreprinted originals
  \item "postings": Postings
  \item "reprints": Reprints
  \item "other": Other
  \item "ads": Ads
}}
}
\value{
A vector of record ids that match the specified reprint status.
}
\description{
This function selects records from a collection based on their reprint status.
It can select records that are reprinted originals, unreprinted originals,
postings, reprints, other or ads.
}
