% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_selecting.R
\name{select_by_season}
\alias{select_by_season}
\title{Selects documents from a collection based on their publication date}
\usage{
select_by_season(
  coll,
  ids = NULL,
  date_MM_DD = NULL,
  days_before = 0,
  days_after = 0
)
}
\arguments{
\item{coll}{A Collection object.}

\item{ids}{A vector of document IDs to select. If NULL, all documents are selected.}

\item{date_MM_DD}{A string indicating the month and day of the year to select documents from. If NULL, all documents are selected.}

\item{days_before}{An integer indicating how many days before the specified date to include in the selection.}

\item{days_after}{An integer indicating how many days after the specified date to include in the selection.}
}
\value{
A Collection object containing only the selected documents.
}
\description{
Selects documents from a collection based on their publication date
}
