% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_selecting.R
\name{select_by_tags}
\alias{select_by_tags}
\title{Selects records from a collection by tags}
\usage{
select_by_tags(
  coll,
  ids = NULL,
  tagslist = NULL,
  headerlist = NULL,
  manualtagslist = NULL
)
}
\arguments{
\item{coll}{A collection object.}

\item{ids}{A vector of record IDs to select. If NULL, all records are selected.}

\item{tagslist}{A vector of tag substrings. Only records containing all of these substrings will be selected.}

\item{headerlist}{A vector of header substrings. Only records containing all of these substrings in their header will be selected.}

\item{manualtagslist}{A vector of manual tag substrings. Only records containing all of these substrings in their manual tags will be selected.}
}
\value{
A vector of record IDs that match the selection criteria.
}
\description{
This function selects records from a collection based on the tags associated with each record.
}
