% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_selecting.R
\name{select_by_text}
\alias{select_by_text}
\title{select_by_text}
\usage{
select_by_text(coll, ids = NULL, searchlist = "")
}
\arguments{
\item{coll}{A collection object.}

\item{ids}{A vector of document ids to search in. If NULL, all documents in the collection are searched.}

\item{searchlist}{A character vector of search terms.}
}
\value{
A vector of document ids that match the search terms.
}
\description{
Selects documents from a collection based on a search list.
}
