% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_showing.R
\name{show_headers}
\alias{show_headers}
\title{Shows headers for all documents in a collection}
\usage{
show_headers(coll)
}
\arguments{
\item{coll}{A collection object.}
}
\value{
A vector of headers associated with all documents in the collection.
}
\description{
This function shows the headers associated with all documents in a collection.
}
