% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_showing.R
\name{show_iiif}
\alias{show_iiif}
\title{Show IIIF images for specified record IDs}
\usage{
show_iiif(ids = NULL, coll = c_all, show_record = TRUE, max.plot = 10)
}
\arguments{
\item{ids}{A vector of record IDs to display.}

\item{coll}{A Collection object.}

\item{show_record}{A logical indicating whether to display record metadata.}

\item{max.plot}{An integer indicating the maximum number of plots to display.}
}
\value{
A plot of IIIF(s) for each record.
}
\description{
This function produces a plot of IIIF(s) for each record.
}
\examples{
show_iiif(ids = c("id1", "id2"), coll = my_collection)
}
