% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_showing.R
\name{show_metadatafields}
\alias{show_metadatafields}
\title{Shows metadata fields for all documents in a collection}
\usage{
show_metadatafields(coll)
}
\arguments{
\item{coll}{A collection object.}
}
\value{
A character vector of metadata fields associated with all documents in the collection.
}
\description{
This function shows the metadata fields associated with all documents in a collection.
}
