% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_showing.R
\name{show_records}
\alias{show_records}
\title{Show Records}
\usage{
show_records(
  coll,
  ids = NULL,
  show_date = TRUE,
  show_text = TRUE,
  show_id = TRUE,
  show_tags = FALSE,
  show_header = FALSE,
  show_position = FALSE
)
}
\arguments{
\item{coll}{Collection object}

\item{ids}{NULL or numeric vector of document ids to show}

\item{show_date}{logical; whether to show date}

\item{show_text}{logical; whether to show text}

\item{show_id}{logical; whether to show id}

\item{show_tags}{logical; whether to show tags}

\item{show_header}{logical; whether to show header}

\item{show_position}{logical; whether to show position}
}
\description{
Show Records
}
