% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_showing.R
\name{show_tags}
\alias{show_tags}
\title{Shows tags for a given set of document IDs}
\usage{
show_tags(coll, ids, manual = FALSE)
}
\arguments{
\item{coll}{A collection object.}

\item{ids}{A vector of document IDs to show tags for.}

\item{manual}{A logical value indicating whether to show manual tags (TRUE) or automatic tags (FALSE).}
}
\value{
A vector of tags associated with the specified documents.
}
\description{
This function shows the tags associated with a given set of document IDs.
}
