% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_visualizing.R
\name{show_wordcloud}
\alias{show_wordcloud}
\title{Convenience Function to Create Quanteda Wordclouds}
\usage{
show_wordcloud(coll, ids = NULL, remove = "", max_words = 200)
}
\arguments{
\item{coll}{A collection object containing the corpus.}

\item{ids}{A vector of document ids to include in the wordcloud.}

\item{remove}{A vector of words to remove from the corpus.}

\item{max_words}{The maximum number of words to include in the wordcloud.}
}
\value{
A wordcloud plot.
}
\description{
This function creates a wordcloud from a corpus of text.
}
\examples{
show_wordcloud(ids = "all", coll = c_all, remove = "", max_words = 200)

}
