% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_things.R
\name{tagfilter_agriculturalobjects}
\alias{tagfilter_agriculturalobjects}
\title{Dictionary Agricultural objects}
\usage{
tagfilter_agriculturalobjects()
}
\description{
Dictionary Agricultural objects
}
