% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_animals.R
\name{tagfilter_animal}
\alias{tagfilter_animal}
\alias{tagfilter_dogs}
\alias{tagfilter_birds}
\alias{tagfilter_cats}
\alias{tagfilter_otherpets}
\alias{tagfilter_livestock}
\alias{tagfilter_animalobjects}
\title{Tagfilter Animals}
\usage{
tagfilter_dogs()

tagfilter_birds()

tagfilter_cats()

tagfilter_otherpets()

tagfilter_livestock()

tagfilter_animalobjects()
}
\description{
Tagfilters are regular expression based filters designed to tag ads in order
to classify ads based on their content. The avisblatt R package comes with
curated filters to search for ads concerning animals, pets (mostly dogs, 
cats, and birds) as well as livestock, and related objects.
}
\details{
The tagfilters help site provides you with a list of available tagfilters
families.
}
\seealso{
tagfilters
}
