% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_animals.R
\name{tagfilter_animalobjects}
\alias{tagfilter_animalobjects}
\title{Tagfilter for objects related to animals}
\usage{
tagfilter_animalobjects()
}
\description{
Tagfilter for objects related to animals
}
