% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_attributes.R
\name{tagfilter_attributes_female}
\alias{tagfilter_attributes_female}
\title{Filter Quanteda Corpus: Transaction type - offers in offer sections}
\usage{
Tagfilters are used internally in the creation of yearly collections of Avisblatt ads, to attribute tags to different ads.
}
\description{
Filter Quanteda Corpus: Transaction type - offers in offer sections
}
\note{
Last changed 2021-11-28
}
\author{
Alexander Engel, ORCID 0000-0002-8592-3124, search terms by Christof Jeggle
}
