% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_main.R
\name{tagfilter_bizpromo}
\alias{tagfilter_bizpromo}
\title{Filter Quanteda Corpus: Advertising/promoting business}
\usage{
tagfilter_bizpromo()
}
\description{
Filter Quanteda Corpus: Advertising/promoting business
}
