% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_work.R
\name{tagfilter_bizpromo1}
\alias{tagfilter_bizpromo1}
\title{Filter Quanteda Corpus: Advertising/promoting business in section labourinfo}
\usage{
tagfilter_bizpromo1()
}
\description{
Filter Quanteda Corpus: Advertising/promoting business in section labourinfo
}
