% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_work.R
\name{tagfilter_bizpromo2}
\alias{tagfilter_bizpromo2}
\title{Filter Quanteda Corpus: Advertising/promoting business outside section labourinfo}
\usage{
tagfilter_bizpromo2()
}
\description{
Filter Quanteda Corpus: Advertising/promoting business outside section labourinfo
}
