% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_work.R
\name{tagfilter_boarding2}
\alias{tagfilter_boarding2}
\title{Filter Quanteda Corpus: Boarding in other sections}
\usage{
tagfilter_boarding2()
}
\description{
Filter Quanteda Corpus: Boarding in other sections
}
