% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_headers.R
\name{tagfilter_boardingheader}
\alias{tagfilter_boardingheader}
\title{Dictionary labour, boarding, and information (Header)}
\usage{
tagfilter_boardingheader()
}
\description{
Dictionary labour, boarding, and information (Header)
}
