% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_main.R
\name{tagfilter_charity1}
\alias{tagfilter_charity1}
\title{Filter Quanteda Corpus: Charity inside charityheader section}
\usage{
tagfilter_charity1()
}
\description{
Filter Quanteda Corpus: Charity inside charityheader section
}
