% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_charity-travel-placement.R
\name{tagfilter_charity2}
\alias{tagfilter_charity2}
\title{Filter Quanteda Corpus: Charity outside charityheader section}
\usage{
tagfilter_charity2()
}
\description{
Filter Quanteda Corpus: Charity outside charityheader section
}
