% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_textiles.R
\name{tagfilter_costume}
\alias{tagfilter_costume}
\title{Dictionary Garments for Special Occassions and Costumes}
\usage{
tagfilter_costume()

tagfilter_costume()
}
\description{
Dictionary Garments for Special Occassions and Costumes

Dictionary Garments for Special Occassions and Costumes
}
