% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_lost.R
\name{tagfilter_crime}
\alias{tagfilter_crime}
\title{Filter Quanteda Corpus: Crime with focus on property delinquency}
\usage{
tagfilter_crime()
}
\description{
Filter Quanteda Corpus: Crime with focus on property delinquency
}
