% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_headers.R
\name{tagfilter_demand}
\alias{tagfilter_demand}
\title{Dictionary Demand (unclear if for sale or to lend) (Header)}
\usage{
tagfilter_demand()
}
\description{
Dictionary Demand (unclear if for sale or to lend) (Header)
}
