% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_headers.R
\name{tagfilter_denaturalisation}
\alias{tagfilter_denaturalisation}
\title{Dictionary Denaturalisation (loss of citizenship) (Header)}
\usage{
tagfilter_denaturalisation()
}
\description{
Dictionary Denaturalisation (loss of citizenship) (Header)
}
