% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_work.R
\name{tagfilter_employment1}
\alias{tagfilter_employment1}
\title{Filter Quanteda Corpus: Employment in labourinfo section}
\usage{
tagfilter_employment1()
}
\description{
Filter Quanteda Corpus: Employment in labourinfo section
}
