% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_work.R
\name{tagfilter_employment2}
\alias{tagfilter_employment2}
\title{Filter Quanteda Corpus: Employment in pertinent sections outside labourinfo}
\usage{
tagfilter_employment2()
}
\description{
Filter Quanteda Corpus: Employment in pertinent sections outside labourinfo
}
