% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_work.R
\name{tagfilter_employmentoffer}
\alias{tagfilter_employmentoffer}
\title{Filter Quanteda Corpus: Employment offer section}
\usage{
tagfilter_employmentoffer()
}
\description{
Filter Quanteda Corpus: Employment offer section
}
