% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_work.R
\name{tagfilter_employmentseeking}
\alias{tagfilter_employmentseeking}
\title{Filter Quanteda Corpus: Employment seeking section}
\usage{
tagfilter_employmentseeking()
}
\description{
Filter Quanteda Corpus: Employment seeking section
}
