% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_things.R
\name{tagfilter_extinguisher}
\alias{tagfilter_extinguisher}
\title{Dictionary Fire Protection and Fire Extinguishers}
\usage{
tagfilter_extinguisher()
}
\description{
Dictionary Fire Protection and Fire Extinguishers
}
