% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_finance.R
\name{tagfilter_finance}
\alias{tagfilter_finance}
\alias{tagfilter_finance_loan1}
\alias{tagfilter_finance_loan2}
\alias{tagfilter_finance_sale1}
\alias{tagfilter_finance_sale2}
\alias{tagfilter_lottery}
\title{Tagfilter Finance}
\usage{
tagfilter_finance_loan1()

tagfilter_finance_loan2()

tagfilter_finance_sale1()

tagfilter_finance_sale2()

tagfilter_lottery()
}
\description{
Tagfilters are regular expression based filters designed to tag ads in order
to classify ads based on their content. The avisblatt R package comes with
curated filters to search for ads related to financial matters such as 
loan requests and offers, trading securities, or advertising class lotteries.
}
\details{
The tagfilters help site provides you with a list of available tagfilters
families.
}
\seealso{
tagfilters
}
