% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_finance.R
\name{tagfilter_finance_loan1}
\alias{tagfilter_finance_loan1}
\title{Filter Quanteda Corpus: Financial loans}
\usage{
tagfilter_finance_loan1()
}
\description{
Filter Quanteda Corpus: Financial loans
}
