% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_finance.R
\name{tagfilter_finance_sale1}
\alias{tagfilter_finance_sale1}
\title{Filter Quanteda Corpus: Financial papers for sale}
\usage{
tagfilter_finance_sale1()
}
\description{
Filter Quanteda Corpus: Financial papers for sale
}
