% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_foodstuff.R
\name{tagfilter_food}
\alias{tagfilter_food}
\alias{tagfilter_spices}
\alias{tagfilter_meat}
\alias{tagfilter_poultry}
\alias{tagfilter_alcohol}
\alias{tagfilter_milk}
\alias{tagfilter_coffee}
\alias{tagfilter_fish}
\alias{tagfilter_fruit}
\alias{tagfilter_driedfruit}
\alias{tagfilter_tropicalfruit}
\alias{tagfilter_nuts}
\alias{tagfilter_honey}
\alias{tagfilter_vegetable}
\alias{tagfilter_mineralwater}
\alias{tagfilter_preserves}
\alias{tagfilter_chocolate}
\alias{tagfilter_grain}
\alias{tagfilter_cheese}
\alias{tagfilter_mushrooms}
\alias{tagfilter_pastry}
\alias{tagfilter_syrup}
\alias{tagfilter_sugar}
\alias{tagfilter_tea}
\alias{tagfilter_oil}
\alias{tagfilter_pasta}
\alias{tagfilter_legumes}
\alias{tagfilter_tobaccoproducts}
\alias{tagfilter_butter}
\alias{tagfilter_eggs}
\alias{tagfilter_broth}
\title{Tagfilter Food}
\usage{
tagfilter_spices()

tagfilter_meat()

tagfilter_poultry()

tagfilter_alcohol()

tagfilter_milk()

tagfilter_coffee()

tagfilter_fish()

tagfilter_fruit()

tagfilter_driedfruit()

tagfilter_tropicalfruit()

tagfilter_nuts()

tagfilter_honey()

tagfilter_vegetable()

tagfilter_mineralwater()

tagfilter_preserves()

tagfilter_chocolate()

tagfilter_grain()

tagfilter_cheese()

tagfilter_mushrooms()

tagfilter_pastry()

tagfilter_syrup()

tagfilter_sugar()

tagfilter_tea()

tagfilter_oil()

tagfilter_pasta()

tagfilter_legumes()

tagfilter_tobaccoproducts()

tagfilter_butter()

tagfilter_eggs()

tagfilter_broth()
}
\description{
Tagfilters are regular expression based filters designed to tag ads in order
to classify ads based on their content. The avisblatt R package comes with
curated filters to search for food and drink related ads and finds ad content from spices,
chocolate and nuts to alcohol, milk and meat.
}
\details{
Tagfilters can only predict if an ad is pertinent to a given topic.
Depending on the complexity of the topic and the development stage of a
tagfilter, there can be a considerable number of false positives and false
negatives.

The tagfilters help site provides you with a list of available tagfilters
families.
}
\seealso{
tagfilters
}
