% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_things.R
\name{tagfilter_health}
\alias{tagfilter_health}
\title{Dictionary Health, Cosmetics and Drugstore Products}
\usage{
tagfilter_health()
}
\description{
Dictionary Health, Cosmetics and Drugstore Products
}
