% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_property.R
\name{tagfilter_housing1}
\alias{tagfilter_housing1}
\title{Filter Quanteda Corpus: Housing within "labourinfo" section}
\usage{
tagfilter_housing1()
}
\description{
Filter Quanteda Corpus: Housing within "labourinfo" section
}
