% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_property.R
\name{tagfilter_housing2}
\alias{tagfilter_housing2}
\title{Filter Quanteda Corpus: Housing outside "labourinfo" section}
\usage{
tagfilter_housing2()
}
\description{
Filter Quanteda Corpus: Housing outside "labourinfo" section
}
