% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_property.R
\name{tagfilter_housing_other1}
\alias{tagfilter_housing_other1}
\title{Filter Quanteda Corpus: Housing offered or requested, unclear if for sale or rent}
\usage{
tagfilter_housing_other1()
}
\description{
Filter Quanteda Corpus: Housing offered or requested, unclear if for sale or rent
}
