% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_property.R
\name{tagfilter_housing_other2}
\alias{tagfilter_housing_other2}
\title{Filter Quanteda Corpus: Housing, unclear if for sale or rent}
\usage{
tagfilter_housing_other2()
}
\description{
Filter Quanteda Corpus: Housing, unclear if for sale or rent
}
