% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_property.R
\name{tagfilter_housing_rent1}
\alias{tagfilter_housing_rent1}
\title{Filter Quanteda Corpus: Housing for rent in renting section}
\usage{
Tagfilters are used internally in the creation of yearly collections of Avisblatt ads, to attribute tags to different ads.
}
\description{
Filter Quanteda Corpus: Housing for rent in renting section
}
\note{
Last changed 2021-08-04
}
\author{
Alexander Engel, ORCID 0000-0002-8592-3124
}
