% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_property.R
\name{tagfilter_housing_rent2}
\alias{tagfilter_housing_rent2}
\title{Filter Quanteda Corpus: Housing for rent in other sections}
\usage{
tagfilter_housing_rent2()
}
\description{
Filter Quanteda Corpus: Housing for rent in other sections
}
