% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_property.R
\name{tagfilter_housing_rent3}
\alias{tagfilter_housing_rent3}
\title{Filter Quanteda Corpus: Housing in conjunction with boarding}
\usage{
tagfilter_housing_rent3()
}
\description{
Filter Quanteda Corpus: Housing in conjunction with boarding
}
