% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_property.R
\name{tagfilter_housing_sale1}
\alias{tagfilter_housing_sale1}
\title{Filter Quanteda Corpus: Housing for sale in sale and auction sections, unambiguous search terms}
\usage{
tagfilter_housing_sale1()
}
\description{
Filter Quanteda Corpus: Housing for sale in sale and auction sections, unambiguous search terms
}
