% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_property.R
\name{tagfilter_housing_sale2}
\alias{tagfilter_housing_sale2}
\title{Filter Quanteda Corpus: Housing for sale in sale and auction sections, more ambiguous search terms}
\usage{
tagfilter_housing_sale2()
}
\description{
Filter Quanteda Corpus: Housing for sale in sale and auction sections, more ambiguous search terms
}
