% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_property.R
\name{tagfilter_housing_sale3}
\alias{tagfilter_housing_sale3}
\title{Filter Quanteda Corpus: Housing for auction sale in the house itself}
\usage{
tagfilter_housing_sale3()
}
\description{
Filter Quanteda Corpus: Housing for auction sale in the house itself
}
