% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_property.R
\name{tagfilter_housing_sale5}
\alias{tagfilter_housing_sale5}
\title{Filter Quanteda Corpus: Housing for sale in other sections}
\usage{
tagfilter_housing_sale5()
}
\description{
Filter Quanteda Corpus: Housing for sale in other sections
}
