% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_titles.R
\name{tagfilter_labourandinfo}
\alias{tagfilter_labourandinfo}
\title{Dictionary labour and information (Title)}
\usage{
tagfilter_labourandinfo()
}
\description{
Dictionary labour and information (Title)
}
