% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_main.R
\name{tagfilter_lostandfound1}
\alias{tagfilter_lostandfound1}
\title{Filter Quanteda Corpus: Lostandfound inside lostandfound section}
\usage{
tagfilter_lostandfound1()
}
\description{
Filter Quanteda Corpus: Lostandfound inside lostandfound section
}
