% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_main.R
\name{tagfilter_lostandfound2}
\alias{tagfilter_lostandfound2}
\title{Filter Quanteda Corpus: Lostandfound outside lostandfound section}
\usage{
tagfilter_lostandfound2()
}
\description{
Filter Quanteda Corpus: Lostandfound outside lostandfound section
}
