% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_headers.R
\name{tagfilter_lostandfoundheader}
\alias{tagfilter_lostandfoundheader}
\title{Dictionary lost and found (Header)}
\usage{
tagfilter_lostandfoundheader()
}
\description{
Dictionary lost and found (Header)
}
