% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_headers.R
\name{tagfilter_merge_to_ad}
\alias{tagfilter_merge_to_ad}
\title{Dictionary of 'false headers' that to do not head a section of ads, but are a heading WITHIN an ad. To be merged to next ad}
\usage{
tagfilter_merge_to_ad()
}
\description{
Dictionary of 'false headers' that to do not head a section of ads, but are a heading WITHIN an ad. To be merged to next ad
}
