% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_household.R
\name{tagfilter_mischousehold}
\alias{tagfilter_mischousehold}
\title{Dictionary Misc Household Goods (Unspecified)}
\usage{
tagfilter_mischousehold()
}
\description{
Dictionary Misc Household Goods (Unspecified)
}
