% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_household.R
\name{tagfilter_mischoushold}
\alias{tagfilter_mischoushold}
\title{Dictionary Misc Household Goods (Unspecified)}
\usage{
tagfilter_mischoushold()
}
\description{
Dictionary Misc Household Goods (Unspecified)
}
