% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_headers.R
\name{tagfilter_naturalisation}
\alias{tagfilter_naturalisation}
\title{Dictionary Naturalisation (gain of citizenship) (Header)}
\usage{
tagfilter_naturalisation()
}
\description{
Dictionary Naturalisation (gain of citizenship) (Header)
}
