% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_headers.R
\name{tagfilter_offer}
\alias{tagfilter_offer}
\title{Dictionary Offer (unclear if for sale or to lend) (Header)}
\usage{
tagfilter_offer()
}
\description{
Dictionary Offer (unclear if for sale or to lend) (Header)
}
