% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_opportunity.R
\name{tagfilter_opportunity}
\alias{tagfilter_opportunity}
\alias{tagfilter_travel1}
\alias{tagfilter_travel2}
\alias{tagfilter_charity1}
\alias{tagfilter_charity2}
\title{Tagfilter opportunity}
\usage{
tagfilter_travel1()

tagfilter_travel2()

tagfilter_charity1()

tagfilter_charity2()
}
\description{
Tagfilters are regular expression based filters designed to tag ads in order
to classify ads based on their content. The avisblatt R package comes with
curated filters to search for ads concerning specific opportunities, like 
joining in a stage ride to another town, or in charitable action.
}
\details{
The tagfilters help site provides you with a list of available tagfilters
families.
}
\seealso{
tagfilters
}
