% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_household.R
\name{tagfilter_plantobject}
\alias{tagfilter_plantobject}
\title{Dictionary Houseplants and Related Objects}
\usage{
tagfilter_plantobject()
}
\description{
Dictionary Houseplants and Related Objects
}
