% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_practices.R
\name{tagfilter_practices}
\alias{tagfilter_practices}
\alias{tagfilter_spinning}
\alias{tagfilter_weaving}
\alias{tagfilter_cooking}
\alias{tagfilter_knitting}
\title{Tagfilter Practices}
\usage{
tagfilter_spinning()

tagfilter_weaving()

tagfilter_cooking()

tagfilter_knitting()
}
\description{
Tagfilters are regular expression based filters designed to tag ads in order
to classify ads based on their content. The avisblatt R package comes with
curated filters to search for practices related ads and finds ads that 
contain different practices from spinning to weaving or cooking.
}
\details{
Tagfilters can only predict if an ad is pertinent to a given topic. 
Depending on the complexity of the topic and the development stage of a 
tagfilter, there can be a considerable number of false positives and false 
negatives. 

The tagfilters help site provides you with a list of available tagfilters
families.
}
\seealso{
tagfilters
}
