% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_print.R
\name{tagfilter_print}
\alias{tagfilter_print}
\alias{tagfilter_print1}
\alias{tagfilter_print2}
\alias{tagfilter_print3}
\alias{tagfilter_print4}
\alias{tagfilter_print5}
\alias{tagfilter_print6}
\alias{tagfilter_print8}
\alias{tagfilter_print9}
\alias{tagfilter_print10}
\title{Tagfilter Print}
\usage{
tagfilter_print1()

tagfilter_print2()

tagfilter_print3()

tagfilter_print4()

tagfilter_print5()

tagfilter_print6()

tagfilter_print8()

tagfilter_print9()

tagfilter_print10()
}
\description{
Tagfilters are regular expression based filters designed to tag ads in order
to classify ads based on their content. The avisblatt R package comes with
curated filters to search for ads concerning books and other printed items,
including promotion by bookstores and publishers, (shared) subscriptions to 
books and journals, and books in specific contexts like books in libraries, 
lost books, or  book lotteries.
}
\details{
The tagfilters help site provides you with a list of available tagfilters
families.
}
\seealso{
tagfilters
}
