% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_main.R
\name{tagfilter_print1}
\alias{tagfilter_print1}
\title{Filter Quanteda Corpus: Print inside bookstore section}
\usage{
tagfilter_print1()
}
\description{
Filter Quanteda Corpus: Print inside bookstore section
}
