% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_print.R
\name{tagfilter_print2}
\alias{tagfilter_print2}
\title{Filter Quanteda Corpus: Print outside bookstore section, commercial sellers}
\usage{
tagfilter_print2()
}
\description{
Filter Quanteda Corpus: Print outside bookstore section, commercial sellers
}
