% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_print.R
\name{tagfilter_print3}
\alias{tagfilter_print3}
\title{Filter Quanteda Corpus: Print outside bookstore section, non-commercial sellers}
\usage{
tagfilter_print3()
}
\description{
Filter Quanteda Corpus: Print outside bookstore section, non-commercial sellers
}
