% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_print.R
\name{tagfilter_print6}
\alias{tagfilter_print6}
\title{Filter Quanteda Corpus: books/journals in libraries}
\usage{
tagfilter_print6()
}
\description{
Filter Quanteda Corpus: books/journals in libraries
}
