% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_property.R
\name{tagfilter_property}
\alias{tagfilter_property}
\alias{tagfilter_housing_rent1}
\alias{tagfilter_housing_rent2}
\alias{tagfilter_housing_sale1}
\alias{tagfilter_housing_sale2}
\alias{tagfilter_housing_sale3}
\alias{tagfilter_housing_sale4}
\alias{tagfilter_housing_sale5}
\alias{tagfilter_housing_other1}
\alias{tagfilter_housing_other2}
\alias{tagfilter_churchseat}
\title{Tagfilter Property}
\usage{
Tagfilters are used internally in the creation of yearly collections of Avisblatt ads, to attribute tags to different ads.

tagfilter_housing_rent2()

tagfilter_housing_sale1()

tagfilter_housing_sale2()

tagfilter_housing_sale3()

tagfilter_housing_sale4()

tagfilter_housing_sale5()

tagfilter_housing_other1()

tagfilter_housing_other2()

tagfilter_churchseat()
}
\description{
Tagfilters are regular expression based filters designed to tag ads in order
to classify ads based on their content. The avisblatt R package comes with
curated filters to search for ads concerning the housing and property market, 
and the (tradable) entitlement to fixed seats in churches.
}
\details{
The tagfilters help site provides you with a list of available tagfilters
families.
}
\note{
Last changed 2021-09-29
}
\seealso{
tagfilters
}
\author{
Alexander Engel, ORCID 0000-0002-8592-3124
}
