% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_headers.R
\name{tagfilter_propertysaleoffer}
\alias{tagfilter_propertysaleoffer}
\title{Dictionary Property Sale Offer (Header)}
\usage{
tagfilter_propertysaleoffer()
}
\description{
Dictionary Property Sale Offer (Header)
}
