% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_titles.R
\name{tagfilter_saledemand}
\alias{tagfilter_saledemand}
\title{Dictionary things demanded for sale (Title)}
\usage{
tagfilter_saledemand()
}
\description{
Dictionary things demanded for sale (Title)
}
