% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_textiles.R
\name{tagfilter_texmaterial}
\alias{tagfilter_texmaterial}
\title{Dictionary Certain Types of Fabric/ Textile Material
This Dictionary is ment to find textiles in ads where no other dictionary is useful,
but a lot of ads are of course already found through other dictionaries;
qualities of textiles will be explored in other dictionaries in more detail}
\usage{
tagfilter_texmaterial()
}
\description{
Dictionary Certain Types of Fabric/ Textile Material
This Dictionary is ment to find textiles in ads where no other dictionary is useful,
but a lot of ads are of course already found through other dictionaries;
qualities of textiles will be explored in other dictionaries in more detail
}
