% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_textiles.R
\name{tagfilter_textiles}
\alias{tagfilter_textiles}
\alias{tagfilter_clothing}
\alias{tagfilter_sleepwear}
\alias{tagfilter_uniform}
\alias{tagfilter_underwear}
\alias{tagfilter_outerwear}
\alias{tagfilter_costume}
\alias{tagfilter_shoes}
\alias{tagfilter_handkerchief}
\alias{tagfilter_hand}
\alias{tagfilter_neck}
\alias{tagfilter_headdress}
\alias{tagfilter_texmaterial}
\alias{tagfilter_cloth}
\alias{tagfilter_yarn}
\title{Tagfilter Property}
\usage{
tagfilter_clothing()

tagfilter_sleepwear()

tagfilter_uniform()

tagfilter_underwear()

tagfilter_outerwear()

tagfilter_costume()

tagfilter_shoes()

tagfilter_handkerchief()

tagfilter_hand()

tagfilter_neck()

tagfilter_headdress()

tagfilter_texmaterial()

tagfilter_cloth()

tagfilter_yarn()
}
\description{
Tagfilters are regular expression based filters designed to tag ads in order
to classify ads based on their content. The avisblatt R package comes with
curated filters to search for textile related ads and finds broader categories like
clothing and mentions of textile materials as well as smaller items like handkerchiefs,
special occasion wear (tag: costume) or muffs and gloves (tag: hand).
}
\details{
Tagfilters can only predict if an ad is pertinent to a given topic. 
Depending on the complexity of the topic and the development stage of a 
tagfilter, there can be a considerable number of false positives and false 
negatives. 
 
The precision and sensitivity of some (families of) tagfilters can be 
measured by comparison to a manual classification for four sample years 
(1734, 1754, 1774 and 1834) from an early stage of the Avisblatt project.
Since the manual classification does often only roughly match the scope of 
the tagfilters, their true precision and sensitivity are underestimated.

Calculated that way, the family of tagfilters concerning textiles shows
a precision >80% and a sensitivity >81%.

The tagfilters help site provides you with a list of available tagfilters
families.
}
\seealso{
tagfilters
}
