% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_things.R
\name{tagfilter_tobaccoobjects}
\alias{tagfilter_tobaccoobjects}
\title{Dictionary Tobacco and Related Objects}
\usage{
tagfilter_tobaccoobjects()
}
\description{
Dictionary Tobacco and Related Objects
}
