% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_transactiontype.R
\name{tagfilter_transactiontype}
\alias{tagfilter_transactiontype}
\alias{tagfilter_transactiontype_offer1}
\alias{tagfilter_transactiontype_request1}
\alias{tagfilter_transactiontype_offer2}
\alias{tagfilter_transactiontype_request2}
\alias{tagfilter_transactiontype_offer3}
\alias{tagfilter_transactiontype_request3}
\title{Tagfilter Transactiontype}
\usage{
Tagfilters are used internally in the creation of yearly collections of Avisblatt ads, to attribute tags to different ads.

Tagfilters are used internally in the creation of yearly collections of Avisblatt ads, to attribute tags to different ads.

Tagfilters are used internally in the creation of yearly collections of Avisblatt ads, to attribute tags to different ads.

Tagfilters are used internally in the creation of yearly collections of Avisblatt ads, to attribute tags to different ads.

Tagfilters are used internally in the creation of yearly collections of Avisblatt ads, to attribute tags to different ads.

Tagfilters are used internally in the creation of yearly collections of Avisblatt ads, to attribute tags to different ads.
}
\description{
#' Tagfilters are regular expression based filters designed to tag ads in 
order to classify ads based on their content. The avisblatt R package comes 
with curated filters to search for ads according to the type of the proposed 
transaction: Is it an offer or a request?
}
\details{
The tagfilters help site provides you with a list of available tagfilters
families.
}
\note{
Last changed 2021-11-03

Last changed 2021-11-03

Last changed 2021-11-03

Last changed 2023-02-01

Last changed 2021-11-03

Last changed 2021-11-03
}
\seealso{
tagfilters
}
\author{
Alexander Engel, ORCID 0000-0002-8592-3124

Alexander Engel, ORCID 0000-0002-8592-3124

Alexander Engel, ORCID 0000-0002-8592-3124

Alexander Engel, ORCID 0000-0002-8592-3124

Alexander Engel, ORCID 0000-0002-8592-3124; most regex by Lars Dickmann

Alexander Engel, ORCID 0000-0002-8592-3124; most regex by Lars Dickmann
}
