% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_transactiontype.R
\name{tagfilter_transactiontype_offer3}
\alias{tagfilter_transactiontype_offer3}
\title{Filter Quanteda Corpus: Transaction type - offers of found items}
\usage{
Tagfilters are used internally in the creation of yearly collections of Avisblatt ads, to attribute tags to different ads.
}
\description{
Filter Quanteda Corpus: Transaction type - offers of found items
}
\note{
Last changed 2021-11-03
}
\author{
Alexander Engel, ORCID 0000-0002-8592-3124; most regex by Lars Dickmann
}
