% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_transactiontype.R
\name{tagfilter_transactiontype_request1}
\alias{tagfilter_transactiontype_request1}
\title{Filter Quanteda Corpus: Transaction type - requests in request/seek/demand sections}
\usage{
Tagfilters are used internally in the creation of yearly collections of Avisblatt ads, to attribute tags to different ads.
}
\description{
Filter Quanteda Corpus: Transaction type - requests in request/seek/demand sections
}
\note{
Last changed 2021-11-03
}
\author{
Alexander Engel, ORCID 0000-0002-8592-3124
}
