% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_headers.R, R/tagfilters_main.R
\name{tagfilter_travel}
\alias{tagfilter_travel}
\title{Dictionary Travel (Header)}
\usage{
tagfilter_travel()

tagfilter_travel()
}
\description{
Dictionary Travel (Header)

Filter Quanteda Corpus: Travel opportunities
}
