% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_main.R
\name{tagfilter_travel1}
\alias{tagfilter_travel1}
\title{Filter Quanteda Corpus: Travel opportunities in travelheader section}
\usage{
tagfilter_travel1()
}
\description{
Filter Quanteda Corpus: Travel opportunities in travelheader section
}
