% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_main.R
\name{tagfilter_travel2}
\alias{tagfilter_travel2}
\title{Filter Quanteda Corpus: Travel opportunities outside travelheader section}
\usage{
tagfilter_travel2()
}
\description{
Filter Quanteda Corpus: Travel opportunities outside travelheader section
}
