% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_things.R
\name{tagfilter_trolley}
\alias{tagfilter_trolley}
\title{Dictionary Trolleys (Carriages for Transportation of Objects, not People)}
\usage{
tagfilter_trolley()
}
\description{
Dictionary Trolleys (Carriages for Transportation of Objects, not People)
}
