% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_things.R
\name{tagfilter_umbrella}
\alias{tagfilter_umbrella}
\title{Dictionary Umbrellas and Related Objects (umbrella cases and umbrella cloth)}
\usage{
tagfilter_umbrella()
}
\description{
Dictionary Umbrellas and Related Objects (umbrella cases and umbrella cloth)
}
