% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_textiles.R
\name{tagfilter_uniform}
\alias{tagfilter_uniform}
\title{Dictionary Military Clothing/ Uniforms}
\usage{
tagfilter_uniform()

tagfilter_uniform()
}
\description{
Dictionary Military Clothing/ Uniforms

Dictionary Military Clothing/ Uniforms
}
