% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_work.R
\name{tagfilter_work}
\alias{tagfilter_work}
\alias{tagfilter_employmentoffer}
\alias{tagfilter_employmentseeking}
\alias{tagfilter_employment1}
\alias{tagfilter_employment2}
\alias{tagfilter_bizpromo1}
\alias{tagfilter_bizpromo2}
\alias{tagfilter_boarding1}
\alias{tagfilter_boarding2}
\title{Tagfilter work}
\usage{
tagfilter_employmentoffer()

tagfilter_employmentseeking()

tagfilter_employment1()

tagfilter_employment2()

tagfilter_bizpromo1()

tagfilter_bizpromo2()

tagfilter_boarding1()

tagfilter_boarding2()
}
\description{
Tagfilters are regular expression based filters designed to tag ads in order
to classify ads based on their content. The avisblatt R package comes with
curated filters to search for ads concerning the job market (especially 
servants, clerks, and apprentices), promoting services, and boarding 
arrangements.
}
\details{
The tagfilters help site provides you with a list of available tagfilters
families.
}
\seealso{
tagfilters
}
