% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilter_doc.R
\name{tagfilters}
\alias{tagfilters}
\title{Avisblatt Regular Expression Based Tagfilters}
\description{
Avisblatt tagfilter functions are parameterless functions
that hold an array of semantically grouped regular expressions to tag, i.e.,
classify ads based on their content. E.g., a dogs tagfilter would contain
a thorough collection of all sorts of regular expressions around dogs
including different historic spelling options.
}
\seealso{
\code{\link{tagfilters_animals}}
}
