% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_main.R
\name{tf_ignorecase}
\alias{tf_ignorecase}
\title{Some tagfilters were build to be case sensitive, some were not. This gives the ones that are not sensitive}
\usage{
tf_ignorecase(prefix = F)
}
\description{
Some tagfilters were build to be case sensitive, some were not. This gives the ones that are not sensitive
}
