% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_collections.R
\name{write_collection}
\alias{write_collection}
\title{Write collection to Disk}
\usage{
write_collection(x, name_on_disk, pretty_json = TRUE, zip = FALSE)
}
\arguments{
\item{x}{object of class Collection.}

\item{name_on_disk}{character name of the file on disk}

\item{pretty_json}{boolean should json be prettified? Defaults to TRUE.}

\item{zip}{boolean should data and data description be put together in one zip archive?
Defaults to FALSE.}
}
\description{
Write a collection to disk using two separate files: a .csv file for the data
and a .json file for the metainformation. Objects are linked across files using
ids.
}
