% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdata_import.R
\name{xml_direct_import}
\alias{xml_direct_import}
\title{Read Transkribus XML Files}
\usage{
xml_direct_import(
  AVIS_YEARS = 1729:1844,
  source_path = "../avis-databuffer/xml",
  dest_path = "../avis-databuffer/raw_data_uncorrected",
  gt_years = c(1734, 1754, 1774, 1834)
)
}
\arguments{
\item{AVIS_YEARS}{numeric vector indicating selected years.}

\item{source_path}{character location of the input file.}

\item{dest_path}{character path to write the data files to.}

\item{gt_years}{numeric vector indicating ground truth years, i.e., that were
manually inspected.}
}
\description{
Read Transkribus XML Files
}
