% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_scouting.R
\name{get_noadverts}
\alias{get_noadverts}
\title{Get noadvert records from a Collection object}
\usage{
get_noadverts(coll, text = FALSE)
}
\arguments{
\item{coll}{A Collection object.}

\item{text}{A logical value indicating whether to return the text of the noadvert records or their names.}
}
\value{
If text is TRUE, the text of the noadvert records; otherwise, their names.
}
\description{
This function returns the noadvert records from a Collection object.
}
\examples{
coll <- read_collection("path/to/collection")
get_noadverts(coll)

}
\seealso{
\code{\link{read_collection}}, \code{\link{gather_collections}}
}
\keyword{collection}
