% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_subcorpus_by_id}
\alias{get_subcorpus_by_id}
\title{Get Subsets of a Corpus by Ids}
\usage{
get_subcorpus_by_id(corp, ids, idvar = "id")
}
\arguments{
\item{corp}{quanteda corpus object}

\item{ids}{character text ids}

\item{idvar}{character name of the id column. Defaults to id.}
}
\description{
Get Subsets of a Corpus by Ids
}
