% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_text_by_id}
\alias{get_text_by_id}
\title{Legacy Function to Get Text by Identifier}
\usage{
get_text_by_id(corp, ids, n = NULL, identifier = "id", txt = "texts")
}
\arguments{
\item{corp}{quanteda corpus object}

\item{ids}{character text ids}

\item{n}{numeric number of hits, defaults to NULL.}

\item{identifier}{character name of id column.}

\item{txt}{character column identifiert. Defaults to 'texts'.}
}
\description{
Legacy Function to Get Text by Identifier
}
