% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_collections.R
\name{merge_collections}
\alias{merge_collections}
\title{Merge One or More Collections Into a One Large Collection}
\usage{
merge_collections(c_list, merge_corpus = TRUE)
}
\arguments{
\item{c_list}{list of collections. Every element needs to be a 'collection'
object from the avisblatt R package.}
}
\description{
Merge One or More Collections Into a One Large Collection
}
