% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_attributes.R
\name{tagfilter_attributes_female}
\alias{tagfilter_attributes_female}
\title{Tagfilter Attributes}
\usage{
tagfilter_attributes_female()
}
\description{
#' Tagfilters are regular expression based filters designed to tag ads in order
to classify ads based on their content. The avisblatt R package comes with
curated filters.
}
\seealso{
tagfilters
}
