% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_finance.R
\name{tagfilter_finance}
\alias{tagfilter_finance}
\alias{tagfilter_finance_loan1}
\alias{tagfilter_finance_loan2}
\alias{tagfilter_finance_sale1}
\alias{tagfilter_finance_sale2}
\alias{tagfilter_lottery}
\title{Tagfilter Finance}
\usage{
tagfilter_finance_loan1()

tagfilter_finance_loan2()

tagfilter_finance_sale1()

tagfilter_finance_sale2()

tagfilter_lottery()
}
\description{
Tagfilters are regular expression based filters designed to tag ads in order
to classify ads based on their content. The avisblatt R package comes with
curated filters to search for ads related to financial matters such as 
loan requests and offers, trading securities, or advertising class lotteries.
}
\details{
Tagfilters can only predict if an ad is pertinent to a given topic. 
Depending on the complexity of the topic and the development stage of a 
tagfilter, there can be a considerable number of false positives and false 
negatives. 

The precision and sensitivity of some (families of) tagfilters can be 
measured by comparison to a manual classification for four sample years 
(1734, 1754, 1774 and 1834) from an early stage of the Avisblatt project.
Since the manual classification does often only roughly match the scope of 
the tagfilters, their true precision and sensitivity are underestimated.

Calculated that way, the family of tagfilters concerning loans and securities 
shows a precision >78% and a sensitivity >64%. The lottery tagfilter shows a 
precision >73% and a sensitivity >95%. 

The tagfilters help site provides you with a list of available tagfilters
families.
}
\seealso{
tagfilters
}
