% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_household.R
\name{tagfilter_household}
\alias{tagfilter_household}
\alias{tagfilter_bed}
\alias{tagfilter_bedding}
\alias{tagfilter_tablelinen}
\alias{tagfilter_carpet}
\alias{tagfilter_chair}
\alias{tagfilter_cabinet}
\alias{tagfilter_stove}
\alias{tagfilter_mirror}
\alias{tagfilter_timepiece}
\alias{tagfilter_table}
\alias{tagfilter_tableware}
\alias{tagfilter_bureau}
\alias{tagfilter_toy}
\alias{tagfilter_game}
\alias{tagfilter_kitchen}
\alias{tagfilter_lighting}
\alias{tagfilter_instrument}
\alias{tagfilter_wallpaper}
\alias{tagfilter_cutlery}
\alias{tagfilter_divider}
\alias{tagfilter_petobject}
\alias{tagfilter_upholstery}
\alias{tagfilter_domestic}
\alias{tagfilter_garden}
\alias{tagfilter_homedeco}
\alias{tagfilter_art}
\alias{tagfilter_bathobject}
\alias{tagfilter_mischousehold}
\title{Tagfilter Household}
\usage{
tagfilter_bed()

tagfilter_bedding()

tagfilter_tablelinen()

tagfilter_carpet()

tagfilter_chair()

tagfilter_cabinet()

tagfilter_stove()

tagfilter_mirror()

tagfilter_timepiece()

tagfilter_table()

tagfilter_tableware()

tagfilter_bureau()

tagfilter_toy()

tagfilter_game()

tagfilter_kitchen()

tagfilter_lighting()

tagfilter_instrument()

tagfilter_wallpaper()

tagfilter_cutlery()

tagfilter_divider()

tagfilter_petobject()

tagfilter_upholstery()

tagfilter_domestic()

tagfilter_garden()

tagfilter_homedeco()

tagfilter_art()

tagfilter_bathobject()

tagfilter_mischousehold()
}
\description{
Tagfilters are regular expression based filters designed to tag ads in order
to classify ads based on their content. The avisblatt R package comes with
curated filters to search for household related ads and finds ad content from beds to
carpets and much much more.
}
\details{
The tagfilters help site provides you with a list of available tagfilters
families.
}
\seealso{
tagfilters
}
