% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_property.R
\name{tagfilter_housing_rent1}
\alias{tagfilter_housing_rent1}
\title{Filter Quanteda Corpus: Housing for rent in renting section}
\usage{
tagfilter_housing_rent1()
}
\description{
Filter Quanteda Corpus: Housing for rent in renting section
}
