% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_lost.R
\name{tagfilter_lostandfound}
\alias{tagfilter_lostandfound}
\alias{tagfilter_lostandfound1}
\alias{tagfilter_lostandfound2}
\alias{tagfilter_crime}
\title{Tagfilter Lost and Found}
\usage{
tagfilter_lostandfound1()

tagfilter_lostandfound2()

tagfilter_crime()
}
\description{
Tagfilters are regular expression based filters designed to tag ads in order
to classify ads based on their content. The avisblatt R package comes with
curated filters to search for ads announcing lost or found, and potentially 
stolen items.
}
\details{
The tagfilters help site provides you with a list of available tagfilters
families.
}
\seealso{
tagfilters
}
