% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_opportunity.R
\name{tagfilter_opportunity}
\alias{tagfilter_opportunity}
\alias{tagfilter_travel1}
\alias{tagfilter_travel2}
\alias{tagfilter_charity1}
\alias{tagfilter_charity2}
\title{Tagfilter opportunity}
\usage{
tagfilter_travel1()

tagfilter_travel2()

tagfilter_charity1()

tagfilter_charity2()
}
\description{
Tagfilters are regular expression based filters designed to tag ads in order
to classify ads based on their content. The avisblatt R package comes with
curated filters to search for ads concerning specific opportunities, like 
joining in a stage ride to another town, or in charitable action.
}
\details{
Tagfilters can only predict if an ad is pertinent to a given topic. 
Depending on the complexity of the topic and the development stage of a 
tagfilter, there can be a considerable number of false positives and false 
negatives. 

The tagfilters help site provides you with a list of available tagfilters
families.
}
\seealso{
tagfilters
}
