% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_print.R
\name{tagfilter_print4}
\alias{tagfilter_print4}
\title{Filter Quanteda Corpus: shared newspaper subscriptions}
\usage{
tagfilter_print4()
}
\description{
Filter Quanteda Corpus: shared newspaper subscriptions
}
