% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagfilters_work.R
\name{tagfilter_work}
\alias{tagfilter_work}
\alias{tagfilter_employmentoffer}
\alias{tagfilter_employmentseeking}
\alias{tagfilter_employment1}
\alias{tagfilter_employment2}
\alias{tagfilter_bizpromo1}
\alias{tagfilter_bizpromo2}
\alias{tagfilter_boarding1}
\alias{tagfilter_boarding2}
\title{Tagfilter work}
\usage{
tagfilter_employmentoffer()

tagfilter_employmentseeking()

tagfilter_employment1()

tagfilter_employment2()

tagfilter_bizpromo1()

tagfilter_bizpromo2()

tagfilter_boarding1()

tagfilter_boarding2()
}
\description{
Tagfilters are regular expression based filters designed to tag ads in order
to classify ads based on their content. The avisblatt R package comes with
curated filters to search for ads concerning the job market (especially 
servants, clerks, and apprentices), promoting services, and boarding 
arrangements.
}
\details{
Tagfilters can only predict if an ad is pertinent to a given topic. 
Depending on the complexity of the topic and the development stage of a 
tagfilter, there can be a considerable number of false positives and false 
negatives. 

The precision and sensitivity of some (families of) tagfilters can be 
measured by comparison to a manual classification for four sample years 
(1734, 1754, 1774 and 1834) from an early stage of the Avisblatt project.
Since the manual classification does often only roughly match the scope of 
the tagfilters, their true precision and sensitivity are underestimated.

Calculated that way, the family of tagfilters concerning employment 
shows a precision >92% and a sensitivity >78%.

The tagfilters help site provides you with a list of available tagfilters
families.
}
\seealso{
tagfilters
}
